+++ Source/Frontend/UI/Input/Input.cs
@@ -30,30 +30,6 @@
 
         readonly HashSet<System.Windows.Forms.Control> WantingMouseFocus = new HashSet<System.Windows.Forms.Control>();
 
-        [Flags]
-        public enum ModifierKeys
-        {
-            // Summary:
-            //     The bitmask to extract modifiers from a key value.
-            Modifiers = -65536,
-
-            // Summary:
-            //     No key pressed.
-            None = 0,
-
-            // Summary:
-            //     The SHIFT modifier key.
-            Shift = 65536,
-
-            // Summary:
-            //     The CTRL modifier key.
-            Control = 131072,
-
-            // Summary:
-            //     The ALT modifier key.
-            Alt = 262144,
-        }
-
         public static Input Instance { get; private set; }
         readonly Thread UpdateThread;
         private bool KillUpdateThread = false;
